Organlzed structure into two folders: include(.hpp files), and src (.cpp files).

+++ penguin_2d/include/vector2.hpp
@@ -5,7 +5,7 @@
 
 namespace Penguin2D {
 
-	template<typename T = float>
+	template<typename T = double>
 
 	struct Vector2 {
